8248186: Move CDS C++ vtable code to cppVtables.cpp
