+++ make/hotspot/lib/JvmFeatures.gmk
@@ -120,6 +120,7 @@ ifneq ($(call check-jvm-feature, cds), true)
       archiveUtils.cpp \
       classListParser.cpp \
       classLoaderExt.cpp \
+      cppVtables.cpp \
       dumpAllocStats.cpp \
       dynamicArchive.cpp \
       filemap.cpp \
